+++ Changelog.md
@@ -8,6 +8,7 @@
 - Fix Catalonia calendar by removing *Sant Juan* day, which does not appear to be an official holiday (#531).
 - Improve coverage of `workalendar/core.py` module (#546).
 - Improve coverage for the Netherlands calendar - Queen's Day (#546).
+- Improve coverage for the Romania calendar - Liberation day (#546).
 
 ## v12.0.0 (2020-10-02)